service/requirements.txt
Flask
requests
google-cloud-bigquery